fix: move components page request adjusts
