+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
-import { Header } from '@/components/Header';
-import { PageWrapper } from '@/components/PageWrapper';
+import { Header } from '@/presentation/components/Header';
+import { PageWrapper } from '@/presentation/components/PageWrapper';
 import { CartProvider } from '@/context/Cart';
 import { ThemeProvider } from '@/context/theme-provider';
 import { Metadata, Viewport } from 'next';
