+++ src/app/shopping/success/layout.tsx
@@ -1,4 +1,4 @@
-import { FeedbackPage } from '@/components/FeedbackPage';
+import { FeedbackPage } from '@/presentation/components/FeedbackPage';
 import successSvg from '@/presentation/assets/images/success.svg';
 
 import shoppingRoutes from '@/flows/shopping/routes';
